app_otel_sim/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
python-dotenv
PyYAML
databricks-sql-connector
databricks-sdk
fastapi
uvicorn
